migrate extension to typescript react with vite
