+++ manifest.json
@@ -8,19 +8,19 @@
     "permissions": ["storage", "scripting", "downloads"],
     "host_permissions": ["https://sora.chatgpt.com/*"],
   
-    "background": { "service_worker": "background.js" },
+    "background": { "service_worker": "src/background.ts" },
   
     "content_scripts": [
       {
         "matches": ["https://sora.chatgpt.com/*"],
-        "js": ["content.js"],
+        "js": ["src/content.tsx"],
         "run_at": "document_start"
       }
     ],
   
     "web_accessible_resources": [
       {
-        "resources": ["pageHook.js"],
+        "resources": ["src/pageHook.ts"],
         "matches": ["https://sora.chatgpt.com/*"]
       }
     ]
